Rename keyv project to kv
