+++ projects/kv/src/main/kotlin/kv/KeyValueStore.kt
@@ -1,4 +1,4 @@
-package keyv
+package kv
 
 interface KeyValueStore {
     /**
