+++ projects/kv/src/main/kotlin/kv/Rocks.kt
@@ -1,4 +1,4 @@
-package keyv
+package kv
 
 import org.rocksdb.InfoLogLevel
 import org.rocksdb.Options
