+++ settings.gradle
@@ -1,7 +1,7 @@
 include "bin"
 include "core"
 include "gps"
-include "keyv"
+include "kv"
 include "log"
 include "microcontrollers"
 include "units"
